# api/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Provides an API for (Android) SMS Gateway",
        "title": "SMS Gateway - API",
        "contact": {
            "name": "Aleksandr Soloshenko",
            "email": "i@capcom.me"
        },
        "version": "{APP_VERSION}"
    },
    "host": "sms.capcom.me",
    "basePath": "/api",
    "paths": {
        "/3rdparty/v1/device": {
            "get": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Returns list of registered devices",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List devices",
                "responses": {
                    "200": {
                        "description": "Device list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smsgateway.Device"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/3rdparty/v1/message": {
            "get": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Returns message state by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Messages"
                ],
                "summary": "Get message state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message state",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MessageState"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Enqueues message for sending. If ID is not specified, it will be generated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Messages"
                ],
                "summary": "Enqueue message",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Skip phone validation",
                        "name": "skipPhoneValidation",
                        "in": "query"
                    },
                    {
                        "description": "Send message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smsgateway.Message"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Message enqueued",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MessageState"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "Get message state URL"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/3rdparty/webhooks": {
            "get": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Returns list of registered webhooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Webhooks"
                ],
                "summary": "List webhooks",
                "responses": {
                    "200": {
                        "description": "Webhook list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Registers webhook. If webhook with same ID already exists, it will be replaced",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Webhooks"
                ],
                "summary": "Register webhook",
                "parameters": [
                    {
                        "description": "Webhook",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO": {
            "type": "object",
            "required": [
                "event",
                "url"
            ],
            "properties": {
                "event": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookEvent"
                        }
                    ],
                    "example": "sms:received"
                },
                "id": {
                    "type": "string",
                    "maxLength": 36,
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/webhook"
                }
            }
        },
        "github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookEvent": {
            "type": "string",
            "enum": [
                "sms:received"
            ],
            "x-enum-varnames": [
                "WebhookEventSmsReceived"
            ]
        },
        "smsgateway.Device": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "Created at (read only)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "deletedAt": {
                    "description": "Deleted at (read only)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "id": {
                    "description": "ID",
                    "type": "string",
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "lastSeen": {
                    "description": "Last seen at (read only)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "name": {
                    "description": "Name",
                    "type": "string",
                    "example": "My Device"
                },
                "updatedAt": {
                    "description": "Updated at (read only)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
        "smsgateway.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "data": {
                    "description": "Error context"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
...
        "smsgateway.Message": {
            "type": "object",
            "required": [
                "message",
                "phoneNumbers"
            ],
            "properties": {
                "id": {
                    "description": "ID (if not set - will be generated)",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "isEncrypted": {
                    "description": "Is encrypted",
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "description": "Content",
                    "type": "string",
                    "maxLength": 65535,
                    "example": "Hello World!"
                },
                "phoneNumbers": {
                    "description": "Recipients (phone numbers)",
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "79990001234"
                    ]
                },
                "simNumber": {
                    "description": "SIM card number (1-3)",
                    "type": "integer",
                    "maximum": 3,
                    "example": 1
                },
                "ttl": {
                    "description": "Time to live in seconds (conflicts with `validUntil`)",
                    "type": "integer",
                    "minimum": 5,
                    "example": 86400
                },
                "validUntil": {
                    "description": "Valid until (conflicts with `ttl`)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "withDeliveryReport": {
                    "description": "With delivery report",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "smsgateway.MessageState": {
            "type": "object",
            "required": [
                "recipients",
                "state"
            ],
            "properties": {
                "id": {
                    "description": "Message ID",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "isEncrypted": {
                    "description": "Encrypted",
                    "type": "boolean",
                    "example": false
                },
                "isHashed": {
                    "description": "Hashed",
                    "type": "boolean",
                    "example": false
                },
                "recipients": {
                    "description": "Recipients states",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/smsgateway.RecipientState"
                    }
                },
                "state": {
                    "description": "State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.ProcessingState"
                        }
                    ],
                    "example": "Pending"
                },
                "states": {
                    "description": "History of states",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "smsgateway.ProcessingState": {
            "type": "string",
            "enum": [
                "Pending",
                "Processed",
                "Sent",
                "Delivered",
                "Failed"
            ],
            "x-enum-comments": {
                "ProcessingStateDelivered": "Delivered",
                "ProcessingStateFailed": "Failed",
                "ProcessingStatePending": "Pending",
                "ProcessingStateProcessed": "Processed (received by device)",
                "ProcessingStateSent": "Sent"
            },
            "x-enum-varnames": [
                "ProcessingStatePending",
                "ProcessingStateProcessed",
                "ProcessingStateSent",
                "ProcessingStateDelivered",
                "ProcessingStateFailed"
            ]
        },
...
        "smsgateway.RecipientState": {
            "type": "object",
            "required": [
                "phoneNumber",
                "state"
            ],
            "properties": {
                "error": {
                    "description": "Error (for `Failed` state)",
                    "type": "string",
                    "example": "timeout"
                },
                "phoneNumber": {
                    "description": "Phone number or first 16 symbols of SHA256 hash",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 10,
                    "example": "79990001234"
                },
                "state": {
                    "description": "State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.ProcessingState"
                        }
                    ],
                    "example": "Pending"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiAuth": {
            "type": "basic"
        },
        "MobileToken": {
            "description": "Mobile device token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}